pyproject.toml
[tool.poetry]
name = "testing_homework"
description = "Testing Homework"
version = "1.0.0"
readme = "README.md"
authors = ["myapp.com"]


[tool.poetry.dependencies]
python = "3.9.15"

django = { version = "^3.2", extras = ["argon2"] }
django-split-settings = "^1.2"
django-axes = "^5.39"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.13"
django-stubs-ext = "^0.7"
django-ratelimit = "^3.0"

psycopg2-binary = "^2.9"
gunicorn = "^20.0"
python-decouple = "^3.6"
structlog = "^22.1"
requests = "^2.28"
attrs = "^22.1"
pydantic = "^1.10"
punq = "^0.6"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.6"
django-querycount = "^0.7"
django-migration-linter = "^4.1"
django-extra-checks = "^0.12"
nplusone = "^1.0"

wemake-python-styleguide = "^0.17"
flake8-pytest-style = "^1.6"
flake8-logging-format = "^0.7"
nitpick = "^0.32"
doc8 = "^1.0"

pytest = "^7.2"
pytest-django = "^4.5"
pytest-cov = "^4.0"
django-coverage-plugin = "^2.0"
covdefaults = "^2.2.0"
pytest-randomly = "^3.12"
pytest-timeout = "^2.1"
hypothesis = "^6.61"
django-test-migrations = "^1.2"

mypy = "^1.0"
django-stubs = "^1.13"
types-requests = "^2.28"

yamllint = "^1.27"
safety = "^2.1"
dotenv-linter = "^0.4"
polint = "^0.4"
dennis = "^1.1"
dump-env = "^1.3"
ipython = "^8.4"
import-linter = "^1.6"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^5.1"
sphinx-autodoc-typehints = "^1.19"
tomli = "^2.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
